refactor: move saved pools rendering
