+++ src/share.js
@@ -22,7 +22,7 @@ if (!lz) {
   lz = mod.default;
 }
 
-const LOCAL_STORAGE_KEY = "costSplitPools";
+export const LOCAL_STORAGE_KEY = "costSplitPools";
 
 // ---- SAVE/LOAD STATE ----
 // Validate the structure and types of the loaded state
@@ -291,71 +291,6 @@ function deletePoolFromLocalStorage(name) {
 }
 
 /**
- * Render a table of saved pools with load and delete actions.
- *
- * @returns {void}
- */
-function renderSavedPoolsTable() {
-  const table = document.getElementById("saved-pools-table");
-  if (!table || typeof localStorage === "undefined") return;
-  table.innerHTML = "";
-
-  const thead = document.createElement("thead");
-  const headerRow = document.createElement("tr");
-  ["Pool", "People", "Transactions", "Actions"].forEach((h) => {
-    const th = document.createElement("th");
-    th.textContent = h;
-    headerRow.appendChild(th);
-  });
-  thead.appendChild(headerRow);
-  table.appendChild(thead);
-
-  const tbody = document.createElement("tbody");
-  let pools;
-  try {
-    const raw = localStorage.getItem(LOCAL_STORAGE_KEY);
-    pools = raw ? JSON.parse(raw) : {};
-  } catch (e) {
-    pools = {};
-  }
-
-  Object.entries(pools).forEach(([name, data]) => {
-    const row = document.createElement("tr");
-    const nameCell = document.createElement("td");
-    nameCell.textContent = name;
-    row.appendChild(nameCell);
-
-    const peopleCell = document.createElement("td");
-    peopleCell.textContent = Array.isArray(data.people)
-      ? data.people.length
-      : 0;
-    row.appendChild(peopleCell);
-
-    const txCell = document.createElement("td");
-    txCell.textContent = Array.isArray(data.transactions)
-      ? data.transactions.length
-      : 0;
-    row.appendChild(txCell);
-
-    const actionsCell = document.createElement("td");
-    const deleteBtn = document.createElement("button");
-    deleteBtn.textContent = "Delete";
-    deleteBtn.addEventListener("click", (e) => {
-      e.stopPropagation();
-      deletePoolFromLocalStorage(name);
-      renderSavedPoolsTable();
-    });
-    actionsCell.appendChild(deleteBtn);
-    row.appendChild(actionsCell);
-
-    row.addEventListener("click", () => loadPoolFromLocalStorage(name));
-    tbody.appendChild(row);
-  });
-
-  table.appendChild(tbody);
-}
-
-/**
  * Trigger a download of the current state as a JSON file.
  */
 function downloadJson() {
@@ -382,7 +317,6 @@ export {
   loadPoolFromLocalStorage,
   listSavedPools,
   deletePoolFromLocalStorage,
-  renderSavedPoolsTable,
   downloadJson,
   validateState,
   applyLoadedState,
